Factor imgix background concerns into a separate service.
Add element-by-element arithmetic routines to PixelSize.
