+++ rover/src/main/kotlin/io/rover/rover/services/assets/AssetService.kt
@@ -9,23 +9,6 @@ import io.rover.rover.services.network.NetworkTask
 import java.net.URL
 import java.util.concurrent.Executor
 
-interface AssetService {
-    /**
-     * Retrieve the needed photo, from caches if possible.
-     *
-     * [completionHandler] will be called on app's main UI thread.
-     *
-     * TODO: retry logic will not exist on the consumer-side of this method, so rather than
-     * NetworkResult, another result<->error optional type should be used instead.
-     *
-     * TODO: no error handling at all, mostly because we don't know what our policy should be yet.
-     */
-    fun getImageByUrl(
-        url: URL,
-        completionHandler: ((NetworkResult<Bitmap>) -> Unit)
-    ): NetworkTask
-}
-
 class AndroidAssetService(
     networkClient: NetworkClient,
     private val ioExecutor: Executor
