--- rover/src/main/kotlin/io/rover/rover/services/assets/SynchronousPipelineStage.kt
@@ -1,13 +1,0 @@
-package io.rover.rover.services.assets
-
-/**
- * A pipeline step that does I/O operations of (fetch, cache, etc.) an asset.
- *
- * Stages will synchronously block the thread while waiting for I/O.
- *
- * Stages should not block to do computation-type work, however; the asset pipeline is
- * run on a thread pool optimized for I/O multiplexing and not computation.
- */
-interface SynchronousPipelineStage<in TInput, out TOutput> {
-    fun request(input: TInput): TOutput
-}
