+++ rover/src/main/kotlin/io/rover/rover/ui/viewmodels/ImageViewModel.kt
@@ -3,9 +3,11 @@ package io.rover.rover.ui.viewmodels
 import android.graphics.Bitmap
 import android.graphics.RectF
 import android.net.Uri
+import android.util.DisplayMetrics
 import io.rover.rover.core.domain.ImageBlock
 import io.rover.rover.core.logging.log
 import io.rover.rover.services.assets.AssetService
+import io.rover.rover.services.assets.ImageOptimizationServiceInterface
 import io.rover.rover.services.network.NetworkResult
 import io.rover.rover.services.network.NetworkTask
 import io.rover.rover.ui.types.PixelSize
@@ -14,11 +16,12 @@ import java.net.URL
 
 class ImageViewModel(
     private val block: ImageBlock,
-    private val assetService: AssetService
+    private val assetService: AssetService,
+    private val imageOptimizationService: ImageOptimizationServiceInterface
 ) : ImageViewModelInterface {
     override fun requestImage(
         targetViewPixelSize: PixelSize,
-        displayDensity: Float,
+        displayMetrics: DisplayMetrics,
         callback: (Bitmap) -> Unit
     ): NetworkTask? {
         val uri = block.image?.url
@@ -26,10 +29,11 @@ class ImageViewModel(
         return if (uri != null) {
             log.v("There is an image to retrieve.  Starting.")
             val uriWithParameters = Uri.parse(uri.toString()).buildUpon().apply {
-                imgixParameters(
+                imageOptimizationService.optimizeImageBlock(
+                    block,
                     targetViewPixelSize,
-                    displayDensity
-                ).forEach { (key, value) ->  appendQueryParameter(key, value) }
+                    displayMetrics
+                )
             }.build()
             val url = URL(uriWithParameters.toString())
 
@@ -61,48 +65,6 @@ class ImageViewModel(
             // the image at its correct aspect ratio given the width
             val heightToWidthRatio = image.height.toFloat() / image.width.toFloat()
             bounds.width() * heightToWidthRatio
-        }
-    }
-
-    /**
-     * We use a service called Imgix to do cloud-side transforms of our images. Here we're using it
-     * for a scale down transform, if needed. On first usage the cloud service will execute the
-     * transform and then cache the result, meaning that all other users on other devices viewing
-     * the same image asset will get the previously processed bits.
-     */
-    private fun imgixParameters(
-        targetViewPixelSize: PixelSize,
-        displayDensity: Float
-    ): Map<String, String> {
-        return if(block.image != null) {
-            // TODO: should any of this Imgix scaling logic be pulled out into a separate concern?
-
-            val imageSizePixels = PixelSize(
-                block.image.width,
-                block.image.height
-            )
-
-            // Now take border width into account.
-            val borderWidth = block.borderWidth.dpAsPx(displayDensity)
-            val targetViewSizeWithoutBorderWidth = PixelSize(
-                targetViewPixelSize.width - borderWidth,
-                targetViewPixelSize.height - borderWidth
-            )
-
-            // if the ultimate image to be rendered on the screen is going to smaller (in terms of
-            // pixel count) than the source, in terms of pixel count (relevant to data plan usage
-            // for users), then we'll ask Imgix to execute the scale operation for us instead.
-            // However, we will ask for an aspect-correct scale from Imgix because we'll end up
-            // scaling up the larger dimension on our end, saving even more bytes.  Note that we
-            // won't need to have View change the scaling mode from the FIT_XY mode we're already
-            // using because no crop operation is going on here.
-            val smallestSize = minOf(imageSizePixels, targetViewSizeWithoutBorderWidth)
-            mapOf(
-                Pair("w", smallestSize.width.toString()),
-                Pair("h", smallestSize.height.toString())
-            )
-        } else {
-            mapOf()
         }
     }
 }
